backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
pandas
joblib
scikit-learn
xgboost
lightgbm